refactor(utils): move meta resolution logic from dom to formatting module
- Move resolveExportMeta from dom.ts to formatting.ts and rename to resolveMeta
- Add TypeScript interfaces ExportMeta and ResolvedExportMeta for type safety
- Update all imports in dom.ts and exportService.ts to use new location
- Fix escaped newline bug in buildDocument function (\\n → \n)
- Improve code organization by centralizing metadata resolution with formatting logic
